+++ src/Data/Monoid/Lens.hs
@@ -13,8 +13,6 @@ module Data.Monoid.Lens
   -- * Monoid Lens Operators
     (<>~), (<<>~)
   , (<>=), (<<>=)
-  -- * Isomorphisms for Monoidal Wrappers
-  , _dual, _endo, _all, _any, _sum, _product, _first, _last
   ) where
 
 import Data.Monoid
@@ -73,61 +71,3 @@ l <<>~ m = l <%~ (`mappend` m)
 (<<>=) :: (MonadState s m, Monoid r) => SimpleLensLike ((,)r) s r -> r -> m r
 l <<>= r = l <%= (`mappend` r)
 {-# INLINE (<<>=) #-}
-
--- | Isomorphism for 'Dual'
-_dual :: Iso a b (Dual a) (Dual b)
-_dual = isos Dual getDual Dual getDual
-{-# INLINE _dual #-}
-
--- | Isomorphism for 'Endo'
-_endo :: Iso (a -> a) (b -> b) (Endo a) (Endo b)
-_endo = isos Endo appEndo Endo appEndo
-{-# INLINE _endo #-}
-
--- | Isomorphism for 'All'
---
--- >>> ala _all foldMap [True,True]
--- True
---
--- >>> ala _all foldMap [True,False]
--- False
-_all :: Simple Iso Bool All
-_all = iso All getAll
-{-# INLINE _all #-}
-
--- | Isomorphism for 'Any'
---
--- >>> ala _any foldMap [False,False]
--- False
---
--- >>> ala _any foldMap [True,False]
--- True
-_any :: Simple Iso Bool Any
-_any = iso Any getAny
-{-# INLINE _any #-}
-
--- | Isomorphism for 'Sum'
---
--- >>> ala _sum foldMap [1,2,3,4]
--- 10
-_sum :: Iso a b (Sum a) (Sum b)
-_sum = isos Sum getSum Sum getSum
-{-# INLINE _sum #-}
-
--- | Isomorphism for 'Product'
---
--- >>> ala _product foldMap [1,2,3,4]
--- 24
-_product :: Iso a b (Product a) (Product b)
-_product = isos Product getProduct Product getProduct
-{-# INLINE _product #-}
-
--- | Isomorphism for 'First'
-_first :: Iso (Maybe a) (Maybe b) (First a) (First b)
-_first = isos First getFirst First getFirst
-{-# INLINE _first #-}
-
--- | Isomorphism for 'Last'
-_last :: Iso (Maybe a) (Maybe b) (Last a) (Last b)
-_last = isos Last getLast Last getLast
-{-# INLINE _last #-}