+++ CHANGELOG.markdown
@@ -11,6 +11,8 @@
 * Added `Control.Lens.Wrapped` providing a canonical isomorphism for newtypes.
 * Repurposed `ala` to be closer to the original design in @newtype@.
 * Added `au` and `alaf`.
+* Removed many isomorphisms for various newtypes. `_const`, `identity`, `_sum`, etc. Use `wrapping Const`, `wrapping Identity`, etc.
+
 
 3.6
 -----